feat(user-interaction): Implement BookmarkControllerTest for GET #879
